pyproject.toml
[tool.poetry]
name = "keymap-drawer"
version = "0.11.0"
description = "A module and CLI tool to help parse and draw keyboard layouts."
authors = ["Cem Aksoylar <caksoylar@gmail.com>"]
readme = ["README.md", "KEYMAP_SPEC.md"]
packages = [{include = "keymap_drawer"}]
include = ["keymap_drawer/py.typed", "resources/zmk_keyboard_layouts.yaml", "resources/qmk_layouts/*.json"]
license = "MIT"
homepage = "https://github.com/caksoylar/keymap-drawer"
repository = "https://github.com/caksoylar/keymap-drawer"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/caksoylar/keymap-drawer/issues"
"Release Notes" = "https://github.com/caksoylar/keymap-drawer/releases"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
pcpp = "^1.30"
pyparsing = "^3.0.9"
pyyaml = "^6.0"
platformdirs = "^3.5.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pylint = "^2.17.4"
black = "^23.3.0"
pylint-pydantic = "^0.1.8"
types-pyyaml = "^6.0.12.10"
deptry = "^0.12.0"

[tool.poetry.group.lsp]
optional = true

[tool.poetry.group.lsp.dependencies]
python-lsp-server = "^1.7.2"

[tool.poetry.scripts]
keymap = 'keymap_drawer.__main__:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
extension-pkg-whitelist = "pydantic"
load-plugins = "pylint_pydantic"

[tool.pylint.'MESSAGES CONTROL']
disable = """
    too-few-public-methods,
"""

[tool.pylint.BASIC]
good-names = "x,y,w,h,r,f,k,p,m,c"
max-line-length = 120

[tool.mypy]
plugins = "pydantic.mypy"

[tool.black]
line-length = 120

[tool.isort]
src_paths = ["keymap_drawer"]
line_length = 120
